refactor: refactor authentication flow for llm providers
- Move authentication logic to requestLLMProviderToken method
- Remove getLLMProviderToken and logout methods from provider classes
- Create types file for openai and sourcegraph providers
